requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
pydantic-settings
python-dotenv
python-multipart
numpy
scikit-learn
SQLAlchemy>=1.4,<2.0
alembic==1.8
asyncpg
psycopg2-binary
PyJWT==2.8
passlib[bcrypt]==1.7
fastapi-limiter==0.1.5
aioredis==2.0
starlette
pytesseract
Pillow
firebase-admin
google-cloud-storage
spacy
transformers
Babel
python-jose[cryptography]
httpx
google-cloud-vision
pytest
pytest-asyncio
pytest-cov

